build: add headless mode for UI integration tests
You can now run UI tests in headless mode.

+++ src/main/java/model/Game.java
@@ -79,7 +79,7 @@ public class Game {
 			}
 		});
 		playerTurn.addListener((observableValue, oldVal, newVal) -> {
-			switch ((int) newVal) {
+			switch (newVal) {
 				case 1:
 					ai1.playTurn();
 					break;
